+++ CRM/CivirulesConditions/Generic/ValueComparison.php
@@ -179,12 +179,7 @@ abstract class CRM_CivirulesConditions_Generic_ValueComparison extends CRM_Civir
   public function isConditionValid(CRM_Civirules_TriggerData_TriggerData $triggerData) {
     $value = $this->getFieldValue($triggerData);
     $compareValue = $this->getComparisonValue();
-    if (empty($value) || empty($compareValue)) {
-     $result = FALSE;
-    }
-    else {
-      $result = $this->compare($value, $compareValue, $this->getOperator());
-    }
+    $result = $this->compare($value, $compareValue, $this->getOperator());
     return $result;
   }